requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
pdfminer.six==20201018
fpdf==1.7.2
transformers==4.27.4
torch==2.0.1
